Mini_test/requirements.txt
Django>=5.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
django-filter
